Refactor pre and post process to use a common helper
All preprocess functions shared the same structure with different mean and std.  As such,
we created a common helper that change the value of the image pixels with the mean and std given.

In the case of the postprocess, we use common functions for remove the common boxes in the different tinyyolos.